+++ crates/cdk/src/cdk_database/mint_memory.rs
@@ -4,15 +4,16 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use tokio::sync::RwLock;
+use tokio::sync::{Mutex, RwLock};
 
 use super::{Error, MintDatabase};
 use crate::dhke::hash_to_curve;
 use crate::mint::{self, MintKeySetInfo, MintQuote};
+use crate::nuts::nut07::State;
 use crate::nuts::{
-    BlindSignature, CurrencyUnit, Id, MeltQuoteState, MintQuoteState, Proof, Proofs, PublicKey,
+    nut07, BlindSignature, CurrencyUnit, Id, MeltQuoteState, MintQuoteState, Proof, Proofs,
+    PublicKey,
 };
-use crate::secret::Secret;
 
 /// Mint Memory Database
 #[derive(Debug, Clone)]
@@ -21,8 +22,8 @@ pub struct MintMemoryDatabase {
     keysets: Arc<RwLock<HashMap<Id, MintKeySetInfo>>>,
     mint_quotes: Arc<RwLock<HashMap<String, MintQuote>>>,
     melt_quotes: Arc<RwLock<HashMap<String, mint::MeltQuote>>>,
-    pending_proofs: Arc<RwLock<HashMap<[u8; 33], Proof>>>,
-    spent_proofs: Arc<RwLock<HashMap<[u8; 33], Proof>>>,
+    proofs: Arc<RwLock<HashMap<[u8; 33], Proof>>>,
+    proof_state: Arc<Mutex<HashMap<[u8; 33], nut07::State>>>,
     blinded_signatures: Arc<RwLock<HashMap<[u8; 33], BlindSignature>>>,
 }
 
@@ -38,6 +39,21 @@ impl MintMemoryDatabase {
         spent_proofs: Proofs,
         blinded_signatures: HashMap<[u8; 33], BlindSignature>,
     ) -> Result<Self, Error> {
+        let mut proofs = HashMap::new();
+        let mut proof_states = HashMap::new();
+
+        for proof in pending_proofs {
+            let y = hash_to_curve(&proof.secret.to_bytes())?.to_bytes();
+            proofs.insert(y, proof);
+            proof_states.insert(y, State::Pending);
+        }
+
+        for proof in spent_proofs {
+            let y = hash_to_curve(&proof.secret.to_bytes())?.to_bytes();
+            proofs.insert(y, proof);
+            proof_states.insert(y, State::Spent);
+        }
+
         Ok(Self {
             active_keysets: Arc::new(RwLock::new(active_keysets)),
             keysets: Arc::new(RwLock::new(
@@ -49,18 +65,8 @@ impl MintMemoryDatabase {
             melt_quotes: Arc::new(RwLock::new(
                 melt_quotes.into_iter().map(|q| (q.id.clone(), q)).collect(),
             )),
-            pending_proofs: Arc::new(RwLock::new(
-                pending_proofs
-                    .into_iter()
-                    .map(|p| (hash_to_curve(&p.secret.to_bytes()).unwrap().to_bytes(), p))
-                    .collect(),
-            )),
-            spent_proofs: Arc::new(RwLock::new(
-                spent_proofs
-                    .into_iter()
-                    .map(|p| (hash_to_curve(&p.secret.to_bytes()).unwrap().to_bytes(), p))
-                    .collect(),
-            )),
+            proofs: Arc::new(RwLock::new(proofs)),
+            proof_state: Arc::new(Mutex::new(proof_states)),
             blinded_signatures: Arc::new(RwLock::new(blinded_signatures)),
         })
     }
@@ -213,21 +219,18 @@ impl MintDatabase for MintMemoryDatabase {
         Ok(())
     }
 
-    async fn add_spent_proofs(&self, spent_proofs: Proofs) -> Result<(), Self::Err> {
-        let mut proofs = self.spent_proofs.write().await;
+    async fn add_proofs(&self, proofs: Proofs) -> Result<(), Self::Err> {
+        let mut db_proofs = self.proofs.write().await;
 
-        for proof in spent_proofs {
+        for proof in proofs {
             let secret_point = hash_to_curve(&proof.secret.to_bytes())?;
-            proofs.insert(secret_point.to_bytes(), proof);
+            db_proofs.insert(secret_point.to_bytes(), proof);
         }
         Ok(())
     }
 
-    async fn get_spent_proofs_by_ys(
-        &self,
-        ys: &[PublicKey],
-    ) -> Result<Vec<Option<Proof>>, Self::Err> {
-        let spent_proofs = self.spent_proofs.read().await;
+    async fn get_proofs_by_ys(&self, ys: &[PublicKey]) -> Result<Vec<Option<Proof>>, Self::Err> {
+        let spent_proofs = self.proofs.read().await;
 
         let mut proofs = Vec::with_capacity(ys.len());
 
@@ -240,41 +243,34 @@ impl MintDatabase for MintMemoryDatabase {
         Ok(proofs)
     }
 
-    async fn add_pending_proofs(&self, pending_proofs: Proofs) -> Result<(), Self::Err> {
-        let mut proofs = self.pending_proofs.write().await;
-
-        for proof in pending_proofs {
-            proofs.insert(hash_to_curve(&proof.secret.to_bytes())?.to_bytes(), proof);
-        }
-        Ok(())
-    }
-
-    async fn get_pending_proofs_by_ys(
+    async fn update_proofs_states(
         &self,
         ys: &[PublicKey],
-    ) -> Result<Vec<Option<Proof>>, Self::Err> {
-        let spent_proofs = self.pending_proofs.read().await;
+        proof_state: State,
+    ) -> Result<Vec<Option<State>>, Self::Err> {
+        let mut proofs_states = self.proof_state.lock().await;
 
-        let mut proofs = Vec::with_capacity(ys.len());
+        let mut states = Vec::new();
 
         for y in ys {
-            let proof = spent_proofs.get(&y.to_bytes()).cloned();
-
-            proofs.push(proof);
+            let state = proofs_states.insert(y.to_bytes(), proof_state);
+            states.push(state);
         }
 
-        Ok(proofs)
+        Ok(states)
     }
 
-    async fn remove_pending_proofs(&self, secrets: Vec<&Secret>) -> Result<(), Self::Err> {
-        let mut proofs = self.pending_proofs.write().await;
+    async fn get_proofs_states(&self, ys: &[PublicKey]) -> Result<Vec<Option<State>>, Self::Err> {
+        let proofs_states = self.proof_state.lock().await;
 
-        for secret in secrets {
-            let secret_point = hash_to_curve(&secret.to_bytes())?;
-            proofs.remove(&secret_point.to_bytes());
+        let mut states = Vec::new();
+
+        for y in ys {
+            let state = proofs_states.get(&y.to_bytes()).cloned();
+            states.push(state);
         }
 
-        Ok(())
+        Ok(states)
     }
 
     async fn add_blind_signatures(